Redid everything. Its now working.
Switched from linux-5.3.5 to linux-5.1.1 and followed online tutorial to the letter. Made sure to make module install correctly and was able to successfully compile linux kernel. After that I was able to test my previous code, which did not work, so I reworked it so now it is able to output random quotes of the day.